requirements.txt
ConfigArgParse==1.2.3
graphviz==0.14.1
awesome-slugify==1.6.5
requests
certifi                # for requests
chardet                # for requests
idna                   # for requests
urllib3                # for requests
katversion
